% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_extent_function.R
\name{BD.extent}
\alias{BD.extent}
\title{Extension of BD framework for phylogenetic and functional beta diversity}
\usage{
BD.extent(Y, dist_spp, nperm = 999, method = "raw")
}
\arguments{
\item{Y}{A matrix of species occurrence or abundance per site.}

\item{dist_spp}{A matrix containing all trait or phylogenetic distances
(upper and lower diagonal elements) among species.}

\item{nperm}{Scalar indicating the number of permutations to be used in taxa and site shuffle null-models.}

\item{method}{Character. If "raw" the BD will be computed using directly matrix Y (with reliable transformation),
if "distance" BD will be computed by using distance matrix.}
}
\value{
A list containing the observed values of BD (BDextend.obs), LCBD (LCBDextend.obs), SCBD (if method
is raw) and the respective values of probability calculating accordingly to site shuffle null models (ptaxa.BDextend, ptaxa.LCBDextend)
}
\description{
Extension of BD framework for phylogenetic and functional beta diversity
}
