% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Camargo_function.R
\name{camargo.eveness}
\alias{camargo.eveness}
\title{Camargo evenness index}
\usage{
camargo.eveness(n_spec, include_zeros = T)
}
\arguments{
\item{n_spec}{numeric vector containing abundances or presence of species in communities}

\item{include_zeros}{Logical, if TRUE (default) the absences are accounted in the calculation of mean}
}
\value{
numeric
}
\description{
Camargo evenness index
}
