% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EE_function.R
\name{EvennessEigen}
\alias{EvennessEigen}
\title{Evenness of Eigenvalues for diversity metrics (EE)}
\usage{
EvennessEigen(
  matrix.M,
  scale = TRUE,
  method = "standardize",
  evenness = "Camargo"
)
}
\arguments{
\item{matrix.M}{A matrix containing diversity metrics values (columns) for each sample (lines)}

\item{scale}{Logical. If TRUE (default) the diversity metrics are scaled to 0 mean and variance of 1 (standardization)}

\item{method}{Character, the method used to perform the scaling of diversity metrics.
Default is "standardize".}

\item{evenness}{Character, the method used to calculate the Evenness of Eigenvalues.
Default is "Camargo". Other options are "Pielou" or "both"}
}
\value{
A scalar (if used "Camargo" or "Pielou") or a numeric vector of length 2
if "both" in evenness argument.
}
\description{
Evenness of Eigenvalues for diversity metrics (EE)
}
