% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportanceVal_function.R
\name{ImportanceVal}
\alias{ImportanceVal}
\title{IV function to compute Importance Values of diversity metrics}
\usage{
ImportanceVal(matrix.M, scale = TRUE, method = "max", stopRule = TRUE)
}
\arguments{
\item{matrix.M}{Matrix containing diversity metrics (columns) values for each sample (lines)}

\item{scale}{Logical. If TRUE (default) the diversity metrics are scaled}

\item{method}{Method used to scale matrix M, default is scaling each diversity metric by their maximmun value}

\item{stopRule}{Logical. If TRUE (default) a stoping rule are used to select only the axis of the PCA that matched the rule (Kaiser-Gutmann criterion)}
}
\value{

}
\description{
IV function to compute Importance Values of diversity metrics
}
